+++ ballast-core/src/commonMain/kotlin/com/copperleaf/ballast/utils.kt
@@ -1,31 +1,14 @@
 package com.copperleaf.ballast
 
+import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.coroutineScope
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.launchIn
+import kotlinx.coroutines.flow.merge
 import kotlinx.coroutines.flow.onEach
 
 /**
- * Observe a Flow of Inputs that will run asynchronously in this MviSideEffectScope. This will
- * allow other Inputs to be processed and not be blocked by this Flow's subscription.
- *
- * The Flow subscription will remain active during the whole life of the associated ViewModel
- * and will be cancelled when the ViewModel is destroyed, when the side-effect is restarted, or
- * it may terminate itself before then when the Flow completes.
- */
-@Suppress("NOTHING_TO_INLINE")
-public suspend inline fun <Inputs : Any, Events : Any, State : Any> SideEffectScope<Inputs, Events, State>.observeFlow(
-    inputs: Flow<Inputs>
-) {
-    coroutineScope {
-        inputs
-            .onEach { postInput(it) }
-            .launchIn(this)
-    }
-}
-
-/**
- * Observe a Flow of Inputs that will run asynchronously in its own MviSideEffectScope. This will
+ * Observe a Flow of Inputs that will run asynchronously in its own [SideEffectScope]. This will
  * allow other Inputs to be processed and not be blocked by this Flow's subscription.
  *
  * The Flow subscription will remain active during the whole life of the associated ViewModel
@@ -35,20 +18,22 @@ public suspend inline fun <Inputs : Any, Events : Any, State : Any> SideEffectSc
  * The side-effect started from the Flow uses the resulting Input's simple class name as
  * its key.
  */
+@ExperimentalCoroutinesApi
 @Suppress("NOTHING_TO_INLINE")
 public inline fun <
     reified Inputs : Any,
     Events : Any,
-    State : Any> InputHandlerScope<Inputs, Events, State>.observeFlow(
-    inputs: Flow<Inputs>,
+    State : Any> InputHandlerScope<Inputs, Events, State>.observeFlows(
+    vararg inputs: Flow<Inputs>,
+    key: String? = Inputs::class.simpleName,
     noinline onRestarted: suspend () -> Unit = { },
 ) {
     sideEffect(
-        key = Inputs::class.simpleName,
+        key = key,
         onRestarted = onRestarted,
     ) {
         coroutineScope {
-            inputs
+            merge(*inputs)
                 .onEach { postInput(it) }
                 .launchIn(this)
         }